feature: refatorando o código de listagem de alunos, na rota gerenciamento
